Fix count_nonzero type on Windows
The result of `count_nonzero` should always be `int64` regardless if one
runs on Unix or Windows. To fix this, we need to convert the type to
`int64` before summing to ensure this constraint is met. Also this
explicit conversion is generally a good idea instead of relying on Dask
to implicitly do the promotion for us.

+++ dask/array/core.py
@@ -2876,7 +2876,7 @@ def where(condition, x=None, y=None):
 
 @wraps(np.count_nonzero)
 def count_nonzero(a, axis=None):
-    return (a != 0).sum(axis=axis)
+    return (a != 0).astype(np.int64).sum(axis=axis)
 
 
 @wraps(np.flatnonzero)
